+++ main/Entities/InterviewIntern.java
@@ -13,12 +13,13 @@ public class InterviewIntern extends Intern{
     /**
      * Construct an interviewIntern, giving them the given interviewInternName,
      * interviewInternAge and interviewInternSkills.
-     *  @param interviewInternName   The name of this Entities.Intern
+     *
+     * @param interviewInternName   The name of this Entities.Intern
      * @param interviewInternAge    The age of this Entities.Intern
      * @param interviewInternSkills A map of skills where the key is the name of the skill and the value is the percentage.
      */
     public InterviewIntern(String interviewInternName, int interviewInternAge,
-                           HashMap<String, Double> interviewInternSkills){
+                           HashMap<String, Integer> interviewInternSkills){
         super(interviewInternName, interviewInternAge, interviewInternSkills);
     }
